lab3/requirements.txt
aiogram==2.11.2
emoji==1.1.0
redis==3.5.3
ujson==4.0.1
uvloop==0.14.0
django==3.2.2
python-telegram-bot==13.5